// app-uruguay-laravel/frontend/src/interfaces/enrollment.interfaces.ts
export interface Enrollment {
    id: number;
    uuid: string;
    name: string;
    created_at?: string;
    updated_at?: string;
    extra_attributes?: Record<string, any>;
}

export interface EnrollmentFormData extends Omit<Enrollment, 'id' | 'created_at' | 'updated_at' | 'extra_attributes'> {
    id?: number;
    photo?: File | null;
    extra_attributes?: Record<string, any>;
}

export interface EnrollmentResponse {
    data: Enrollment | Enrollment[];
    message?: string;
    success: boolean;
}